+++ packages/graphql-playground-react/src/components/Playground/TopBar/Reload.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import ReloadIcon from './ReloadIcon'
+
+export interface Props {
+  isReloadingSchema: boolean
+  onReloadSchema?: () => void
+}
+
+const Reload: React.SFC<Props> = props => (
+  <ReloadIcon
+    animate={props.isReloadingSchema}
+    onClick={props.onReloadSchema}
+  />
+)
+
+export default Reload
